pyproject.toml
[project]
name = "exiftree"
version = "0.1.0"
description = "Browse photography through the gear that made it."
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.85",
    "celery>=5.6.3",
    "django>=6.0.4",
    "django-storages>=1.14.6",
    "exifread>=3.5.1",
    "pillow>=12.2.0",
    "redis>=7.4.0",
]
